update code location
